vxc_adv_visualizer/requirements.txt
# VXC/ADV Flow Measurement System Dependencies

# Core hardware and data
pyserial>=3.5
h5py>=3.0
numpy>=1.20
pyyaml>=5.4

# GUI and visualization
PyQt5>=5.15
pyqtgraph>=0.12
matplotlib>=3.3

# Optional: Data analysis
scipy>=1.7
pandas>=1.2

# Development tools
pytest>=6.0
